extract xml export to standalone code
make xml generation a two stage process:
1. create a canonical JS representation of the team object
2. emit that JS representation as JS

Additionally, use a namespaced (v5) uuid instead of a random one (v4).  The team name + skater name id results in the same uuid generated for a given skater identifier.  This allows an export to be stable if it is performed multiple times, ensuring that dupes are not added when the xml file is imported.